requirements.txt
aiomysql==0.1.1
alembic==1.8.1
anyio==3.6.2
asyncio==3.4.3
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.18.0
exceptiongroup==1.0.4
fastapi==0.87.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.1
httpx==0.23.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
Mako==1.2.4
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.2
python-decouple==3.8
python-jose==3.3.0
python-slugify==8.0.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.6
starlette==0.21.0
tomli==2.0.1
trio==0.22.0
typer==0.7.0
typing_extensions==4.4.0
uvicorn==0.20.0
zipp==3.10.0
